refactor: split main.rs and app.rs
